+++ ui/src/telegrafs/components/CollectorCard.tsx
@@ -13,7 +13,7 @@ import InlineLabels from 'src/shared/components/inlineLabels/InlineLabels'
 import {
   addTelegrafLabelsAsync,
   removeTelegrafLabelsAsync,
-} from 'src/telegrafs/actions'
+} from 'src/telegrafs/actions/thunks'
 import {createLabel as createLabelAsync} from 'src/labels/actions'
 
 // Selectors